Uses light resource instead of group resource
The group resource is according to phillips, less performant. It only makes sense to use that resource when there are a dozen or more lights associated with that group. Since I'm only dealing with 3 lights, making multiple requests with the lights resource appears to be the better option, and the program will maintain the grouping for me.